pyproject.toml
[tool.poetry]
name = "irys"
version = "0.1.0"
description = "Intelligent Repository System"
authors = ["IRySTeam <padocon@naver.com>"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.8.1"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.9"}
PyJWT = "^2.4.0"
uvicorn = "^0.18.3"
fastapi = "^0.95"
celery = "^5.2.7"
gunicorn = "^20.1.0"
fastapi-event = "^0.1.3"
pythondi = "^1.2.4"
ujson = "^5.4.0"
aiomysql = "^0.1.1"
click = "^8.1.3"
greenlet = "1.1.3"
redis = "^4.3.4"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"
elasticsearch = "^8.6.2"
aiohttp = "^3.8.4"
python-multipart = "^0.0.6"
tika = "^2.6.0"
pytesseract = "^0.3.10"
bcrypt = "^4.0.1"
python-semantic-release = "^7.33.2"
flower = "^1.2.0"
pdf2image = "^1.16.3"
numpy = "^1.24.2"
opencv-python = "^4.7.0.72"
nltk = "^3.8.1"
pandas = "^1.5.3"
scikit-learn = "^1.2.2"
python-magic = "^0.4.27"
asgiref = "^3.6.0"
bert-serving-client = "^1.10.0"
fastapi-mail = "^1.2.6"
uvloop = "^0.17.0"
fitz = "^0.0.1.dev2"
pymupdf = "^1.21.1"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
